+++ src/ll/api/event/player/PlayerRespawnEvent.cpp
@@ -1,13 +1,12 @@
 #include "ll/api/event/player/PlayerRespawnEvent.h"
 #include "ll/api/event/Emitter.h"
 #include "ll/api/memory/Hook.h"
-#include "mc/world/actor/provider/ActorAttribute.h"
 
 namespace ll::event::inline player {
 
 LL_TYPE_INSTANCE_HOOK(PlayerRespawnEventHook, HookPriority::Normal, Player, &Player::$respawn, void) {
     // If the player returns from end or just joins server, his health will > 0
-    if (ActorAttribute::getHealth(getEntityContext()) <= 0) {
+    if (getHealth() <= 0) {
         EventBus::getInstance().publish(PlayerRespawnEvent(*this));
     }
     origin();