+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Removed `Player::getCertificate`
+
 ## [1.8.0-rc.2] - 2026-01-17
 
 ### Changed